fix(ci): pin the toolchain, clear clippy, stop double-running jobs
Clippy passed locally but failed on GitHub because the clippy workflow
installed no toolchain at all — it linted with whatever the runner
shipped, while development ran on nightly. rust-toolchain.toml now pins
an exact stable (the workspace no longer needs nightly; only the fuzzers
do), with clippy/rustfmt and the wasm target, and the workflow installs
it.

The lint CI tripped on, result_large_err, fires inside auto-lsp's
configure_parser expansion (ParseError, ~144 bytes) — not boxable
without forking auto-lsp, so ast allows it with a note. The seven real
warnings a cold cache exposed in db and mir are fixed properly.

ci-clippy/ci-rust/ci-stdlib push triggers are limited to main: an open
pull request used to run every job twice (push + pull_request).

+++ src/tests/lsp/denormalize.rs
@@ -111,7 +111,7 @@ fn denormalize_position_to_offset_round_trip(
     let lsp = denormalize(&with_db, file, &raw).unwrap();
     let offset = position_to_offset(&with_db, file, lsp.start).unwrap();
 
-    assert_eq!(offset, raw.start_byte as usize);
+    assert_eq!(offset, raw.start_byte);
 }
 
 /// A multi-byte char on an earlier line must not shift columns on later lines (line-local conversion).
